test(testkit-js): consolidate events e2e suites to cut CI cost
Folds core/shielded coverage into the provider read-surface suite and
trims proof-heavy variants whose decode paths are covered elsewhere:
2 CI jobs instead of 4, 7 proofs instead of 12.
